+++ .github/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 Update <small>_ May 2024</small>
 
-- feat: add support for autocomplete (22/05/2024)
+- feat: add support for autocomplete (23/05/2024)
+- feat: migrate birthday handler to scheduler job (23/05/2024)
+- chore: load .env before config (23/05/2024)
 - fix: lint workflow failing silently (18/05/2024)
 - chore: fix eol issues on Windows (17/05/2024)